+++ pkg-manager/core/src/install/index.ts
@@ -1094,7 +1094,7 @@ const _installInContext: InstallFunction = async (projects, ctx, opts) => {
         dedupeDirectDeps: opts.dedupeDirectDeps,
         dependenciesByProjectId,
         depsStateCache,
-        disableRelinkFromStore: opts.disableRelinkFromStore,
+        disableRelinkLocalDirDeps: opts.disableRelinkLocalDirDeps,
         extraNodePaths: ctx.extraNodePaths,
         force: opts.force,
         hoistedDependencies: ctx.hoistedDependencies,