+++ store/store-controller-types/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Minor Changes
 
-- 03cdccc6e: New option added: disableRelinkFromStore.
+- 03cdccc6e: New option added: disableRelinkLocalDirDeps.
 
 ### Patch Changes